Godot 4: migrated Betauer.GameTools + tests. Still pending screen services and background loader

+++ Betauer.GameTools/Application/MainLoopNotificationsHandler.cs
@@ -17,7 +17,7 @@ namespace Betauer.Application {
         public event Action OnPostinitialize;
         public event Action OnPredelete;
 
-        public void Execute(int what) {
+        public void Execute(long what) {
             switch (what) {
                 case MainLoop.NotificationOsMemoryWarning:
                     OnOsMemoryWarning?.Invoke();
